LUCENE-3289: add options to FST Builder to tradeoff RAM/CPU used during build vs how small the resulting FST is
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_3x@1145309 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/contrib/misc/src/java/org/apache/lucene/index/MultiPassIndexSplitter.java
@@ -195,7 +195,7 @@ public class MultiPassIndexSplitter {
 
     @Override
     public int numDocs() {
-      return in.maxDoc() - (int)dels.cardinality();
+      return in.maxDoc() - dels.cardinality();
     }
 
     /**
